Код был поделен на отдельные модули
